Make completer smarter.
Suggest node labels when the last non-alpha is a colon and the last
parentheses is closer than the last bracket. Conversely, suggest rel
types when the last non-alpha is a colon and the last bracket is closer
than the last parentheses. Suggest node properties if the last
non-alpha is a period.