Don't suggest properties for non-identifiers
Properties were being suggested anytime the most recent non-alpha was a
period. We only want to suggest properties when the character preceding
the period is an alphabetic character. This makes it so the user
doesn’t get property suggestions when they don’t need it, e.g. when
requesting variable-length paths:

()-[8..9]->()

or when typing floats:

WHERE n.prop > 5.2

Fixes #8.