Slice all the lists in print_schema
The schema was being altered while manipulated in the print_schema
method. Slicing the lists creates a copy so that the original is
untouched. This was affecting the autocompletions after the fact, where
the empty strings that were being inserted to extend all columns to the
same length were being shown in the autocompletions menu.

Fixes #28.