Add the read only check to the Cypher class and fix bugs
The check for whether or not the query was a write query was too naive;
it would abort queries that had the stop words even if they were node
labels, rel types, etc. I’ve also moved the method off into the Cypher
class, as it makes the most sense to me there. Also added tests.

Fixes #29.